File_Type_Identification/requirements.txt.txt
python-magic-bin
colorama
